refactor: move plan types to domain/plan.go, document placeholder behavior
- Move MemberPlan, TerminalPlan, WorkspacePlan, FileEntry, PromptSnapshot,
  EnvSnapshot from domain/team.go to domain/plan.go
- Document that Session.Plan retains {{CLIER_*}} placeholders
