+++ internal/domain/session.go
@@ -18,12 +18,14 @@ const (
 // Session is an execution instance of a Team on a local machine.
 // Plan is built fresh at session start from the team's current state.
 type Session struct {
-	ID        string        `json:"id"`
-	TeamID    string        `json:"team_id"`
-	Status    SessionStatus `json:"status"`
-	Plan      []MemberPlan  `json:"plan"`
-	CreatedAt time.Time     `json:"created_at"`
-	StoppedAt *time.Time    `json:"stopped_at"`
+	ID     string        `json:"id"`
+	TeamID string        `json:"team_id"`
+	Status SessionStatus `json:"status"`
+	// Plan retains {{CLIER_*}} placeholders as built. Safe for name/ID lookups;
+	// paths and commands require resolution before use.
+	Plan      []MemberPlan `json:"plan"`
+	CreatedAt time.Time    `json:"created_at"`
+	StoppedAt *time.Time   `json:"stopped_at"`
 }
 
 func NewSession(id, teamID string) (*Session, error) {
